requirements.txt
tkinter
Pillow
scipy
numpy
opencv-python
matplotlib
pandas